Bug fixes when running on asic

+++ tofino/libsycon/include/sycon/args.h
@@ -11,8 +11,8 @@ struct args_t {
   bool run_ucli;
   int tna_version;
   time_ms_t expiration_time;
-  int in_port;
-  int out_port;
+  uint16_t in_port;
+  uint16_t out_port;
   bool model;
 
   // Wait until the input and output ports are ready.
